- change file_size of file, vendor_compliance_reports from int to int64

+++ pkg/server/api/console/v1/types/types.go
@@ -2047,7 +2047,7 @@ type VendorBusinessAssociateAgreement struct {
 	ValidUntil *time.Time `json:"validUntil,omitempty"`
 	FileName   string     `json:"fileName"`
 	FileURL    string     `json:"fileUrl"`
-	FileSize   int        `json:"fileSize"`
+	FileSize   int64      `json:"fileSize"`
 	CreatedAt  time.Time  `json:"createdAt"`
 	UpdatedAt  time.Time  `json:"updatedAt"`
 }
@@ -2062,7 +2062,7 @@ type VendorComplianceReport struct {
 	ValidUntil *time.Time `json:"validUntil,omitempty"`
 	ReportName string     `json:"reportName"`
 	FileURL    string     `json:"fileUrl"`
-	FileSize   int        `json:"fileSize"`
+	FileSize   int64      `json:"fileSize"`
 	CreatedAt  time.Time  `json:"createdAt"`
 	UpdatedAt  time.Time  `json:"updatedAt"`
 }
@@ -2111,7 +2111,7 @@ type VendorDataPrivacyAgreement struct {
 	ValidUntil *time.Time `json:"validUntil,omitempty"`
 	FileName   string     `json:"fileName"`
 	FileURL    string     `json:"fileUrl"`
-	FileSize   int        `json:"fileSize"`
+	FileSize   int64      `json:"fileSize"`
 	CreatedAt  time.Time  `json:"createdAt"`
 	UpdatedAt  time.Time  `json:"updatedAt"`
 }
